pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pygithub>=2.0.0",
    "gitpython>=3.1.0",
    "markupsafe>=2.1.0",
    "werkzeug>=3.0.0",
]
